requirements.txt
fastapi
uvicorn
pydantic-settings
pymongo
